components/miscellaneous/pipeline-prerequisites/requirements.txt
pyyaml
requests
dask[distributed]